Initial groups implementation
This provides an initial implementation of user groups; design is
described here:

https://docs.google.com/document/d/1aIjHNzWSw3Y1Zm3ndEjmkV4yjL2hUlxaQJo_aGI_dgE/edit#.

It contains initial implementations of the following features, which
I'll be testing and cleaning up in coming commits.

- UserGroup class
- At least one distribution function (bonus for one member from each
class)
- Add UserGroup field to User class
- Create /group and /ungroup chat commands
- User messaging when player is added / removed from group
- Group whispering
- Cleanup group on user disconnect
- Make sure group gets deleted when all members leave

Group members are not being highlighted (still needs work) and I need to
do a good bit of testing for the features mentioned above.