fix for SERVER-124 when user walks out of tile
If a User walks into a tile, then drops a few items, everything is added
to the QuadTreeNode's objects list.  If the user walks out of the tile,
the server removes object[0] from the list (the User), and was
additionally re-sorting some of the remaining items.  fixed that by
simply removing the item.