Initial monsters spawns, with lots of hard-coded defaults
Replaces Monolith with a ControlMessage-based spawn facility that reads
mobs from XML. In a bit more detail:
- Mobs are now being read from XML files (in new monsters/ subdir), and
spawns are being read in from maps. Most of the monster stats are
currently being ignored aside from name, min # of spawns, and max # of
spawns.
- MonsterSpawnJob periodically checks all maps and queue up spawn tasks
for maps that spawn monsters.
- Each spawn task spawns monsters at random locations across the map and
notifies all users on the map of their location.